+++ src/util/util.ts
@@ -923,3 +923,15 @@ export type TileJSON = {
     center?: [number, number, number];
     vector_layers: [{id: string}]; // this is partial but enough for what we need
 };
+
+/**
+ * The maximum world tile zoom (Z).
+ * In other words, the upper bound supported for tile zoom.
+ */
+export const MAX_TILE_ZOOM = 25;
+
+/**
+ * The minimum world tile zoom (Z).
+ * In other words, the lower bound supported for tile zoom.
+ */
+export const MIN_TILE_ZOOM = 0;